Migrate views into individual files, clean up URLs
