+++ cassdegrees/templates/createsubplan.html
@@ -10,7 +10,7 @@
         <p class="{% if is_error %}msg-error{% else %}msg-success{% endif %}">{{ msg }}</p>
     {% endif %}
 
-    <form class="anuform" action="/create_subplan/" method="post">
+    <form class="anuform" action="" method="post">
 
         {% csrf_token %}
