+++ cassdegrees/templates/manageprograms.html
@@ -11,7 +11,7 @@
     {% endif %}
 
     <div id="{{ action }}Page" {% if action == 'Delete' or render.hide_form %}hidden{% endif %}>
-        <form class="anuform" id="mainForm" action="/create_program/" novalidate method="post">
+        <form class="anuform" id="mainForm" action="" novalidate method="post">
             {% csrf_token %}
 
             <fieldset>
